Add inlining CSS/JS in HTML template, or automatically with --entrypoint
This allows for a user to pass a truthy "inline" param and have the CSS
and JS files be embedded into the HTML template rather than loaded
externally. This value is automatically set if "--entrypoint" is
supplied.

This now allows users to output a single HTML file via --entrypoint with
no external dependencies whatsoever.