Add support for MQTT server config generation
Fixed generation of the MQTT config file; part of this change also
specifies that the template parameter "server" must be specified
(similar to the nodejs generator). If the AsyncApi spec file has
multiple servers, this allows the generation step to find the right one
to use.

Also centralizes handler method naming for consistency.