fix: change broken link pointing to semver syntax
In the generator docs for the configuration generator, the link
pointing to the semver syntax is broken. The correct link
(https://docs.npmjs.com/cli/v6/using-npm/semver) which this broken
link should point to is now version dependent. This means it might
soon become broken with future version updates.

Change the link pointing to the semver syntax, to a beginner-friendly
semver article hosted on the learn section of the nodejs website.

Add a read more link that points to a summarised semver article
hosted on the npm website; which contains more resources and a video
demo.

fixes issue #691